--- Namo_SwiftUI/Projects/Feature/Calendar/Sources/FeatureCalendarSourcesTemplate.swift
@@ -1,7 +1,0 @@
-//
-//  emptyfile.swift
-//  Manifests
-//
-//  Created by 정현우 on 10/1/24.
-//
-
