--- Namo_SwiftUI/Projects/Feature/Home/Sources/Components/CalendarItem.swift
@@ -1,1 +1,0 @@
-
